[Issue 79] Supress 64-bit gcc warnings.  Here needed to modify df1b2fun.h file.  The sizes for pointers address' should be size_t and not int.

+++ admb/trunk/src/df1b2-separable/df1b2fun.h
@@ -436,7 +436,7 @@ void print_derivatives(df1b2_header * px,const char * s,
     smartlist(unsigned int bufsize,const adstring& filename);
     double report_usage(void) 
     { 
-      return double(int(bptr)-int(buffer))/double(int(buffend)-int(buffer));
+      return double(size_t(bptr)-size_t(buffer))/double(size_t(buffend)-size_t(buffer));
     }
   };
     
@@ -485,7 +485,7 @@ void print_derivatives(df1b2_header * px,const char * s,
     void operator += (int);
     double report_usage(void) 
     { 
-      return double(int(bptr)-int(buffer))/double(int(buffend)-int(buffer));
+      return double(size_t(bptr)-size_t(buffer))/double(size_t(buffend)-size_t(buffer));
     }
     void write(void * p,int n);
     void write(int n);
@@ -554,7 +554,7 @@ void print_derivatives(df1b2_header * px,const char * s,
     void operator += (int);
     double report_usage(void) 
     { 
-      return double(int(bptr)-int(buffer))/double(int(buffend)-int(buffer));
+      return double(size_t(bptr)-size_t(buffer))/double(size_t(buffend)-size_t(buffer));
     }
     void write(void * p,int n);
     void write(int n);
@@ -616,7 +616,7 @@ void print_derivatives(df1b2_header * px,const char * s,
     void operator += (int);
     double report_usage(void) 
     { 
-      return double(int(bptr)-int(buffer))/double(int(buffend)-int(buffer));
+      return double(size_t(bptr)-size_t(buffer))/double(size_t(buffend)-size_t(buffer));
     }
     void write(void * p,int n);
     void write(int n);
